4-PACS-Module/Orthanc/medical-reporting-module/api/realtime_api.py
#!/usr/bin/env python3
"""
Real-time API for Medical Reporting Module
Handles real-time transcription and live updates
"""

from flask import Blueprint, request, jsonify
from flask_socketio import emit, join_room, leave_room
import logging
from datetime import datetime
import uuid

logger = logging.getLogger(__name__)

realtime_bp = Blueprint('realtime', __name__)


@realtime_bp.route('/health', methods=['GET'])
def health_check():
    """Health check endpoint"""
    return jsonify({
        'status': 'healthy',
        'service': 'realtime_api',
        'timestamp': datetime.utcnow().isoformat()
    })


@realtime_bp.route('/session', methods=['POST'])
def create_session():
    """Create a new real-time session"""
    try:
        data = request.get_json()
        user_id = data.get('user_id', 'demo_user') if data else 'demo_user'
        
        session_id = str(uuid.uuid4())
        
        session = {
            'session_id': session_id,
            'user_id': user_id,
            'created_at': datetime.utcnow().isoformat(),
            'status': 'active',
            'type': 'transcription'
        }
        
        return jsonify(session), 201
        
    except Exception as e:
        logger.error(f"Error creating real-time session: {e}")
        return jsonify({'error': 'Failed to create session'}), 500


@realtime_bp.route('/session/<session_id>', methods=['GET'])
def get_session(session_id):
    """Get session information"""
    try:
        # Mock session data - in real implementation, get from database
        session = {
            'session_id': session_id,
            'user_id': 'demo_user',
            'created_at': datetime.utcnow().isoformat(),
            'status': 'active',
            'type': 'transcription',
            'chunks_processed': 0,
            'total_duration': 0
        }
        
        return jsonify(session)
        
    except Exception as e:
        logger.error(f"Error getting session {session_id}: {e}")
        return jsonify({'error': 'Session not found'}), 404


@realtime_bp.route('/session/<session_id>/close', methods=['POST'])
def close_session(session_id):
    """Close a real-time session"""
    try:
        # Mock session closure
        return jsonify({
            'session_id': session_id,
            'status': 'closed',
            'message': 'Session closed successfully'
        })
        
    except Exception as e:
        logger.error(f"Error closing session {session_id}: {e}")
        return jsonify({'error': 'Failed to close session'}), 500


@realtime_bp.route('/transcribe/chunk', methods=['POST'])
def transcribe_chunk():
    """Process audio chunk for real-time transcription"""
    try:
        # Check if audio file is present
        if 'audio' not in request.files:
            return jsonify({'error': 'No audio file provided'}), 400
        
        audio_file = request.files['audio']
        session_id = request.form.get('session_id')
        chunk_id = request.form.get('chunk_id', str(uuid.uuid4()))
        
        if not session_id:
            return jsonify({'error': 'Session ID required'}), 400
        
        # Mock transcription - in real implementation, use Whisper
        mock_transcription = "This is a mock real-time transcription result. "
        
        result = {
            'session_id': session_id,
            'chunk_id': chunk_id,
            'transcription': mock_transcription,
            'confidence': 0.95,
            'processing_time': 0.5,
            'timestamp': datetime.utcnow().isoformat()
        }
        
        return jsonify(result)
        
    except Exception as e:
        logger.error(f"Error processing chunk: {e}")
        return jsonify({'error': 'Failed to process audio chunk'}), 500


@realtime_bp.route('/status', methods=['GET'])
def get_realtime_status():
    """Get real-time service status"""
    try:
        status = {
            'service': 'realtime_api',
            'status': 'operational',
            'active_sessions': 0,
            'processing_queue': 0,
            'average_latency': 0.8,
            'uptime': '99.9%'
        }
        
        return jsonify(status)
        
    except Exception as e:
        logger.error(f"Error getting real-time status: {e}")
        return jsonify({'error': 'Failed to get status'}), 500


# SocketIO event handlers for real-time communication
from core.app_factory import socketio

@socketio.on('join_session')
def on_join_session(data):
    """Handle client joining a real-time session"""
    try:
        session_id = data.get('session_id')
        user_id = data.get('user_id', 'demo_user')
        
        if session_id:
            join_room(session_id)
            emit('session_joined', {
                'session_id': session_id,
                'user_id': user_id,
                'status': 'connected'
            })
            logger.info(f"User {user_id} joined session {session_id}")
        
    except Exception as e:
        logger.error(f"Error joining session: {e}")
        emit('error', {'message': 'Failed to join session'})


@socketio.on('leave_session')
def on_leave_session(data):
    """Handle client leaving a real-time session"""
    try:
        session_id = data.get('session_id')
        user_id = data.get('user_id', 'demo_user')
        
        if session_id:
            leave_room(session_id)
            emit('session_left', {
                'session_id': session_id,
                'user_id': user_id,
                'status': 'disconnected'
            })
            logger.info(f"User {user_id} left session {session_id}")
        
    except Exception as e:
        logger.error(f"Error leaving session: {e}")
        emit('error', {'message': 'Failed to leave session'})


@socketio.on('audio_chunk')
def on_audio_chunk(data):
    """Handle real-time audio chunk processing"""
    try:
        session_id = data.get('session_id')
        chunk_data = data.get('chunk_data')
        chunk_id = data.get('chunk_id', str(uuid.uuid4()))
        
        if not session_id or not chunk_data:
            emit('error', {'message': 'Invalid chunk data'})
            return
        
        # Mock processing - in real implementation, process with Whisper
        result = {
            'session_id': session_id,
            'chunk_id': chunk_id,
            'transcription': 'Real-time transcription result',
            'confidence': 0.92,
            'timestamp': datetime.utcnow().isoformat()
        }
        
        # Emit result back to the session room
        emit('transcription_result', result, room=session_id)
        
    except Exception as e:
        logger.error(f"Error processing audio chunk: {e}")
        emit('error', {'message': 'Failed to process audio chunk'})


@realtime_bp.errorhandler(404)
def not_found(error):
    return jsonify({'error': 'Real-time resource not found'}), 404


@realtime_bp.errorhandler(500)
def internal_error(error):
    return jsonify({'error': 'Real-time service error'}), 500