4-PACS-Module/Orthanc/medical-reporting-module/requirements.txt
# Ubuntu Patient Care - Requirements
# Medical AI Platform Dependencies

# Core Flask Application
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-CORS==4.0.0

# Database and ORM
SQLAlchemy==2.0.23
Alembic==1.12.1

# AI and Machine Learning
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1
transformers==4.35.2
scipy==1.11.4
numpy==1.24.4

# Audio Processing
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# Medical Data Processing
pydicom==2.4.3
Pillow==10.1.0
opencv-python==4.8.1.78

# Web and API
requests==2.31.0
urllib3==2.0.7
Werkzeug==2.3.7

# Security and Encryption
cryptography==41.0.7
bcrypt==4.0.1
PyJWT==2.8.0

# Data Validation and Serialization
marshmallow==3.20.1
WTForms==3.1.0

# Configuration and Environment
python-dotenv==1.0.0
configparser==6.0.0

# Development and Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0

# Medical Compliance and Audit
audit-log==1.2.0
compliance-tools==2.1.0

# Performance and Monitoring
gunicorn==21.2.0
gevent==23.9.1

# Utilities
click==8.1.7
python-dateutil==2.8.2
pytz==2023.3

# Medical-specific libraries
medical-imaging==1.5.2
healthcare-utils==2.3.1

# Development Dependencies (install separately with pip install -r requirements-dev.txt)
# pytest-mock==3.12.0
# sphinx==7.2.6
# sphinx-rtd-theme==1.3.0
